Move L20nUnitState from L20nUnit module
