+++ mozilla_pootle/static/js/editor/formats/l20n/L20nUnit.js
@@ -8,17 +8,8 @@
 
 import { FTLASTParser, FTLASTSerializer, getPluralForms } from 'l20n';
 
-
-/*
-    Tests can be run in plugin app environment where there is no installed
-    dependencies from the main Pootle. Import fake `t` for tests.
-*/
-let t;
-try {
-  t = require('pootle/shared/utils/i18n').t;
-} catch (e) {
-  t = (value, ctx) => (interpolate(value, ctx, true));
-}
+import { L20nUnitSimpleState, L20nUnitPluralsState,
+         L20nUnitTraitsState, L20nUnitRawState } from './L20nUnitState';
 
 
 function getEmptyTextElementL20nEntity(value) {
@@ -26,191 +17,6 @@ function getEmptyTextElementL20nEntity(value) {
   textElementResource[0].body[0].value.elements[0].value = value;
 
   return textElementResource[0].body[0];
-}
-
-
-class L20nUnitState {
-  constructor(l20nUnit) {
-    this.l20nUnit = l20nUnit;
-    this.values = [];
-  }
-}
-
-
-class L20nUnitSimpleState extends L20nUnitState {
-  constructor(l20nUnit) {
-    super(l20nUnit);
-    this.values = [this.l20nUnit.entity.value.source];
-  }
-
-  getEditorState() {
-    return {
-      values: this.values,
-      hasPlurals: false,
-      isRichModeEnabled: false,
-    }
-  }
-
-  setEmptyEntity(localeCode) {
-    this.l20nUnit.entity.value.elements[0].value = '';
-    this.values = [''];
-  }
-
-  update(i, value) {
-    // equivalent this.values = [value];
-    this.values[i] = value;
-    this.l20nUnit.entity.value.elements[0].value = value;
-  }
-}
-
-
-class L20nUnitPluralsState extends L20nUnitState {
-  constructor(l20nUnit) {
-    super(l20nUnit);
-
-    const variants = this.l20nUnit.entity.value.elements[0].expressions[0].variants;
-    this.values = [];
-    this.pluralForms = [];
-
-    for (let i = 0; i < variants.length; i++) {
-      this.values.push(variants[i].value.source);
-      this.pluralForms.push(this.extractPluralFormName(variants[i]));
-    }
-
-  }
-
-  extractPluralFormName(variant) {
-    const key = FTLASTSerializer.dumpExpression(variant.key);
-    if (variant.default) {
-      return `${key}, default`;
-    }
-    return key;
-  }
-
-  getEditorState() {
-    return {
-      values: this.values,
-      hasPlurals: true,
-      isRichModeEnabled: false,
-    }
-  }
-
-  getPluralFormName(index) {
-    if (this.pluralForms !== undefined && index < this.pluralForms.length) {
-      return t('Plural form [%(key)s]', { key: this.pluralForms[index] });
-    }
-  }
-
-  getShortPluralFormName(index) {
-    if (this.pluralForms !== undefined && index < this.pluralForms.length) {
-      return `[${this.pluralForms[index]}]`;
-    }
-  }
-
-  setEmptyEntity(localeCode) {
-    this.pluralForms = gatherPluralForms(localeCode, this.pluralForms);
-    const resource = FTLASTParser.parseResource(`unit = ${getEmptySelectorPattern(this.pluralForms)}`);
-    this.l20nUnit.entity = resource[0].body[0];
-    this.values = new Array(this.pluralForms.length).fill('');
-  }
-
-  update(i, value) {
-    this.values[i] = value;
-    if (value !== '') {
-      const variantL20nUnit = new L20nUnit(value);
-      this.l20nUnit.entity.value.elements[0].expressions[0].variants[i].value = variantL20nUnit.entity.value;
-    } else {
-      //TODO: mark non-full state
-    }
-  }
-
-  addVariant(pluralForm) {}
-
-  setDefault(i) {}
-}
-
-
-class L20nUnitTraitsState extends L20nUnitState {
-  constructor(l20nUnit) {
-    super(l20nUnit);
-
-    const traits = this.l20nUnit.entity.traits;
-    this.values = [];
-    this.traitLabels = [];
-    for (let i = 0; i < traits.length; i++) {
-      this.values.push(traits[i].value.source);
-      let key = FTLASTSerializer.dumpExpression(traits[i].key);
-      this.traitLabels.push(key);
-    }
-  }
-
-  getEditorState() {
-    return {
-      values: this.values,
-      hasPlurals: true,
-      isRichModeEnabled: false,
-    }
-  }
-
-  getPluralFormName(index) {
-    if (this.traitLabels !== undefined) {
-      return `[${this.traitLabels[index]}]`;
-    }
-  }
-
-  getShortPluralFormName(index) {
-    return this.getPluralFormName(index);
-  }
-
-  setEmptyEntity(localeCode) {
-    const resource = FTLASTParser.parseResource(`unit = ${getEmptyTraitsPattern(this.traitLabels)}`);
-    this.l20nUnit.entity = resource[0].body[0];
-    this.values = new Array(this.traitLabels.length).fill('');
-  }
-
-  update(i, value) {
-    this.values[i] = value;
-    if (value !== '') {
-      const variantL20nUnit = new L20nUnit(value);
-      this.l20nUnit.entity.traits[i].value.elements[0].value = variantL20nUnit.entity.value;
-    } else {
-      //TODO: mark non-full state
-    }
-
-  }
-}
-
-
-class L20nUnitRawState extends L20nUnitState {
-  constructor(l20nUnit) {
-    super(l20nUnit);
-    this.values = [l20nUnit.value];
-  }
-
-  getEditorState() {
-    return {
-      values: this.values,
-      hasPlurals: false,
-      isRichModeEnabled: true,
-    }
-  }
-
-  setEmptyEntity(localeCode) {
-    this.l20nUnit.entity.value.elements[0].value = '';
-    // TODO: check if the next line is necessary here
-    this.l20nUnit.value = '';
-    this.values = [''];
-  }
-
-  update(i, value) {
-    // equivalent this.values = [value];
-    this.values[i] = value;
-
-    const updatedL20nUnit = new L20nUnit(value);
-    if (updatedL20nUnit.entity !== null) {
-      this.l20nUnit.entity = updatedL20nUnit.entity;
-    }
-  }
 }
